requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
python-dotenv
fastapi
uvicorn
pydantic
jinja2
python-multipart
plotly 
shap 
sqlalchemy
mysql-connector-python
python-jose
passlib[bcrypt]
python-dotenv
shap 


